pyproject.toml
[project]
name = "bookstore"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django-filter>=24.3",
    "django>=5.1.4",
    "djangorestframework>=3.15.2",
    "drf-yasg>=1.21.8",
]
